Splitting app in folders
