+++ src/api/get-product.js
@@ -1,0 +1,16 @@
+import config from '../config'
+
+import { getToken } from './get-token'
+
+export const getProducts = async () => {
+  const { access_token } = await getToken()
+  const res = await fetch(`${config.apiUrl}/api/skus`, {
+    headers: {
+      'Content-type': 'application/json',
+      Authorization: `Bearer ${access_token}`,
+    },
+  })
+
+  const json = await res.json()
+  return json
+}
